Fix linters and update readme file
