+++ README.md
@@ -1,2 +1,36 @@
-# rb-template
-A ruby template including Rubcop and Rspec
+# An OOP school library app: Add Person, Student and Teacher classes 🧑‍🏫
+
+> Adopting an Object-oriented programming model to build a simple school library application
+
+## Getting Started
+
+This repository includes files with Ruby that can be used to test the created main classes from the app
+
+Clone this link into your local device.
+
+## Authors
+
+👤 **Jose**
+
+- GitHub: [@jr-cast](https://github.com/jr-cast)
+- Twitter: [@josercastanos](https://twitter.com/josercastanos)
+- LinkedIn: [@jr-cast](https://linkedin.com/in/jr-cast)
+
+## 🤝 Contributing
+
+Contributions, issues, and feature requests are welcome!
+
+
+## Show your support
+
+Give a ⭐️ if you like this project!
+
+## Acknowledgments
+
+- Hat tip to anyone whose code was used
+- Inspiration
+- etc
+
+## 📝 License
+
+This project is [MIT](./MIT.md) licensed.
