+++ classes/student.rb
@@ -1,13 +1,12 @@
-require "./person.rb"
+require './person'
 
 class Student < Person
   def initialize(classroom)
-    super(name, age = "Unknown", parent_permission = true)
+    super(name, age, parent_permission: true)
     @classroom = classroom
-  end 
+  end
 
   def play_hooky
     "¯\(ツ)/¯"
-  end 
+  end
 end
-
